# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PPI Api",
        "version": "0.1"
    },
    "paths": {
...
        "/api/translation/store": {
            "post": {
                "tags": [
                    "Translations"
                ],
                "summary": "Create new Translation",
                "description": "Create new Translation",
                "operationId": "createTranslation",
                "requestBody": {
                    "description": "Translation to store",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "title": "id",
                                        "description": "id of Translation",
                                        "type": "integer",
                                        "readOnly": "true",
                                        "example": "1"
                                    },
                                    "name": {
                                        "title": "name",
                                        "description": "name of Translation",
                                        "type": "string",
                                        "example": "Error 404 page"
                                    },
                                    "render": {
                                        "title": "render",
                                        "description": "render Text",
                                        "type": "string",
                                        "example": "Bohužel, stránka se nepodařila načíst."
                                    },
                                    "language_id": {
                                        "title": "language_id",
                                        "description": "id of Language",
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Translation stored",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Translation"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation exception"
                    }
                }
            }
        },
        "/api/translation/get": {
            "get": {
                "tags": [
                    "Translations"
                ],
                "summary": "Get all Translations",
                "description": "Get all Translation",
                "operationId": "getTranslations",
                "requestBody": {},
                "responses": {
                    "200": {
                        "description": "Get Translations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Translation"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation exception"
                    }
                }
            }
        },
        "/api/translation/patch/{id}": {
            "patch": {
                "tags": [
                    "Translations"
                ],
                "summary": "Update Translation",
                "description": "Update Translation",
                "operationId": "patchTranslation",
                "parameters": [
                    {
                        "parameter": "id",
                        "name": "id",
                        "in": "path",
                        "description": "The ID to patch",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Translation to store",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "title": "name",
                                        "description": "name of Translation",
                                        "type": "string",
                                        "example": "Error 404 page"
                                    },
                                    "render": {
                                        "title": "render",
                                        "description": "render Text",
                                        "type": "string",
                                        "example": "Bohužel, stránka se nepodařila načíst."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Translation patched",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Translation"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation exception"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Translation": {
                "title": "Translation",
                "description": "Translation model",
                "required": [],
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id of Translation",
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "name": {
                        "title": "name",
                        "description": "name of Translation",
                        "type": "string",
                        "example": "Error 404 page"
                    },
                    "render": {
                        "title": "render",
                        "description": "what will be Rendered",
                        "type": "string"
                    },
                    "language_id": {
                        "title": "language_id",
                        "description": "id of Language",
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object"
            },
...
        },
...
    }
}